Replaced CartItem and CheckoutItem classes with common ProductItem class
